+++ changelog.markdown
@@ -112,4 +112,5 @@
 * Corrected the markdown-syntax in this file.
 * Reverted to calling "**error**" rather than "**Control.Monad.fail**", since the **String**-argument for the latter is discarded in **Monad**-implementations other than **IO**.
 * Uploaded to [GitHub](https://github.com/functionalley/ToolShed.git).
-
+* Simplified file **src-test/Main.hs**.
+* Relaxed test **prop_shuffleDistribution** in **ToolShed.Test.System.Random**.